Song one is just four notes in 4/4 time.
Song two is the same but plays a chord on the third beat. The chord is perfectly sychronized on the 3rd beat.

Song three is the same as two but with an extra note on the third beat one tick behind the chord. This is to test that that note is still picked up after processing the chord together as part of the block.