Use channel 10 for percussion
This enables the generation of “percussion” as an “instrument” in the
soundfont builder.  It also adds that instrument to our General Midi
database, with ID 0 (one-based) resp. -1 (zero-based).  This does not
really reflect the use in GM, but avoids a large number of case
distinctions and appears mostly beneficial.  This instrument is
pre-selected for channel 10 (one-based), to conform with GM usage.